test(api): close db connection when done with all the tests
While the database connection should automatically close when the user-driven event loop is done or when the db goes out of scope, this does not seem to be implemented currently in Seqeulize. It is marked as an issue on their repository, but for the time being we have to manually close the connection. The sequelize team would like to fix this, but there is an underlying issue, described [here](https://github.com/sequelize/sequelize/issues/8468#issuecomment-420897880)