ci: publish master to staging.appsemble.com
The job only runs on the master branch in our own repository. Unlike
review apps, the staging environment is deployed automatically. The
environment name is `review/master`, because this matches the pattern
used for our Kubernetes integration.

refs #101.

The issue should not be closed before it has been validated the
deployment on master is succesful.