Make navigation type configurable
Multiple navigation types are now possible. `bottom` is now an
alternative to the default side menu.

It is now possible to specify an icon on a page. The icon will be
rendered in the menu.

Also a reusable `<Icon />` component was created and is now implemented
throughout the code base.