Add support for SimpleForm field validation
The validation in based on HTML5 validation.

A help message is always displayed to prevent jumping of lower form
elements when the user is typing.

For consistency, snly one of error message or help message is displayed,
not both.

Meaningful error messages are added to the forms that currently use
SimpleForm.