Build the server using Babel
For now, the server is built using the Babel CLI. This is a step towards
the usage of TypeScript for the backend.

The runtime dependencies for the server are compiled using `@babel/cli`.
This removes the need for ESM in production, and allows the usage of
TypeScript for production code. If a package is fully converted, the
`prepack` script can be converted to use `tsc`, just like in
`@appsemble/sdk`.

The CLI and `create-appsemble` now use `@babel/register` at runtime
instead of ESM. This has a negative impact on performance, but this is a
temporary solution and shouldn’t be much of an issue.

The Docker image and Helm chart have been updated not to use ESM
anymore.

As a proof of concept, most of `@appsemble/utils` has been converted to
TypeScript.