Fix how old browsers are handled
Instead of using a special entry point, the script is inlined. A CSP
nonce is required to allow this script to run. Also the wrappers of the
fallback content have been changed from `<div>` to
`<script type="text/x-template">`. This should prevent browsers from
interpreting the content and downloading the images used in this
fallback.

The inline JavaScript is not checked using ESLint, but it is formatted
using Prettier. This is desired behaviour, because syntax is required to
support older browsers, which would make ESLint complain.