Run Lighthouse in for staging apps
This runs Lighthouse for apps in the staging environment. This is done
programmatically, so the process can exit with exit code 1 if baselines
are not met.

App names have been changed to match the directory they are in. This is
because the new Lighthouse script assumes the directory name matches the
url node.

Lighthouse reports are output to the reports directory, and are exposed
as artifacts.

`@appsemble/scripts` is a place for more future automated tasks.

Very minimal Lighthouse typings have been added, as the state of typings
within the Lighthouse repository is a bit weird.

Puppeteer 1.x was used, because one of our dependencies already uses
this. Installing multiple puppeteer dependencies, means installing
multiple Chrome binaries, which is a lot of data.