Flatten publish block API
The publich block API now accepts a flat `BlockDefinition` in its API,
instead of one nested in the property `data`. Instead of
`patternProperties`, the files are now uploaded as a `files` array.

The result is that the `BlockDefinition` schema can be reused for the
request and response bodies, and other binary files can be added to the
API.

To achieve this, the filenames have to be URL encoded.

This is based on !975.

This is a prerequisite for #338.