Add useConfirmation hook
This hook returns a function. If this function is called, the user is
prompted to confirm the action. This replaces custom modal logic for
simple prompts in several components.

Also a small fix was made in `<ClienCredentials />`, so the submit
button is removed when a user has created new client credentials.