Implement sequelize-typescript
This enables us to use TypeScript decorators to define models. This
removes the need to duplicate model fields and use their TypeScript
properties to define how database tables are structured.

The goal of this change is not to cleanup the database structure, but to
synchronize the models with the database as-is. This means some unwanted
constructs have been ported as well.

Some of the constraints have slightly changed. This is ok, since the
database wasn’t fully compatible with the database currently in
production anyway.

The database has been compared by exporting the database before and
after the model changes using the following command:

```sh
pg_dump --schema-only -U admin -h localhost appsemble --no-owner --no-comments
```

The order of the fields has changed, so they need to be sorted manually
to see the diff properly.