Make most jobs interruptible
Jobs that don’t trigger deployments are now interruptible. This means
jobs these jobs will be cancelled if the pipeline is replaced because of
another commit to a merge request.

This fixes the issue that the provision and e2e steps of an older
pipeline caused the newer pipeline to fail if they happen to be
provisioning a newer environment.