Rewrite reducer based async loops to for of
Using `await` inside a for loop is disallowed by
`eslint-config-airbnb-base`. To work around this, a reducer was used.
The goal of this rule is to decrease performance hits when the code is
transpiled by Babel for legacy browsers. Since we don’t target these
browsers, this rule has been disabled inside our codebase.

Using `for of` loops is clearer than the reducer syntax. It’s easy to
forget to await the reducer.

Closes #405