Implement ingress wilcard URLs and certificates
The previous way of mapping URLs in Kubernetes is discarded in favor of
the following method:

For each organization an ingress is created. The ingress maps a wildcard
DNS rule to the service and registers a wildcard certificate.

For each app that has a custom domain name, an ingress is registered as
well.

Now that multiple ingresses are registered, the URL mapping is handled
by another resource than one managed by Helm. This means Helm will no
longer clean up the URL mapping on new deployments.

The downside is that the cleanup needs to be done by the Appsemble
deployment. This is done in a Helm `pre-uninstall` hook, since this
requires the registered clusterrole to exist.