Remove usage of isValidElement in preact-components
This was added to sync up with the React version, with the main
difference being that errors in React can be expected to be managed
error messages that are valid elements.

In preact-components, these are more often than not user-defined
strings, which aren't elements created by h/createElement.

This causes the function to always return false and not display error
messages.