Add a reusable webpack config
This allows block developers to just start coding without the need to
manage a webpack configuration file themselves.

It’s still possible for block developers to override the configuration.

The package also includes type definitions for the Webpack
configuration.

The CLI has been updated to fallback to this package by default if no
`webpack.config` file can be resolved in the block directory.