Reorganize Appsemble Studio page components
Appsemble Studio now has a pages folder. The pages folder defines all
routes. Each page can be defined in one of these categories: routes,
page, or index page.

Each folder in the pages directory maps to a URL. If the URL is a
variable, the name is contextual.

Each routes component exposed a React component of the name `*Routes`.
This component renders a `<Switch />` or `<MetaSwitch />`.

Each routes component contains one index route in the `Index/index.tsx`
file. The exported component is named `IndexPage`.

Other pages are named after the URL they map to, postfixed with `Page`.