Only show the max length on inputs if needed
This only shows the length indicator on inputs if the user approaches
the size limits. Most fields should have `maxLength` set, but for many
fields this is just to catch edge cases most users won’t encounter.

The indicitor is now shown if the text size exceeds the 80% of the
character limit.