Move @types dependencies to their respective dependants
If a package isn’t published, there’s no need to make a distinction
between `dependencies` and `devDependencies`, so all dependencies  may be
stored in `dependencies`.

If a package is published and the built package depends on types of a
dependency, the types belong in `dependencies`, otherwise they belong in
`devDependencies`.

If a package doesn’t expose a JavaScript API, i.e. the server and CLI,
the types aren’t needed in production and belong in `devDependencies`.