Fix exceptional SAML login issue
If an app had logged into an account using SAML before the app had a
default role, and the default role has been added later, the user was
unable to login to the app again.

This is now resolved by assigning the default role to the user if they
don’t have a role yet, even if they had previously logged in to an
earlier version of the app without a default role.