Introduce readData and writeData
These functions support reading and writing data from JSON or YAML
files. When writing files, keys are sorted and the content is formatted
using Prettier.

This fixes an error in the CLI when it tries to load a YAML file as JSON
in some situations.

The function is also reused by the release script.

The translation files have been verified, which sorted them
alphabetically.