Replace oneOf JSON schemas with anyOf
A `oneOf` JSON schema means it must match exactly one of the given
schemas. An `anyOf` property means it may match one or more of the given
schemas.

Because of our strict schemas, the result is the same. However, because
`oneOf` must validate the value against all schemas, and `anyOf` only
needs to find one match, using `anyOf` is more performant.