Remove OAuth2Consent table
It has been replaced with a `consent` column in the `AppMember` table.

At the moment, the API will still expect scopes to be sent and received, this can be removed if necessary.

Also, if this is merged before the release of `0.18.28`, the migration for `0.18.29` should be merged with the one for `0.18.28`.