Provide consistent app validation in the API
Several custom checks are applied on an app definition. Instead of
throwing immediately if one is found, all violations are collected and
thrown as one single validation error. This all errors can be
communicated once, so errors can be resolved better.

The validator of the `jsonschema` package is used to collect errors.
Custom checks are used to append custom validation errors. Because the
error format is consistent, it can be used by automated tooling to
provide better feedback to the user.

Refs #641