Replace js-yaml with yaml
The `yaml` package allows us to work with YAML on an AST level,
`js-yaml` doesn’t.

A prerelease version was chosen to match the version which is used by an
upcoming release of `monaco-yaml`. This prevents the need to pull in
different versions of `yaml` and prevents refactors in the future.

This change simply replaces `js-yaml` everywhere. Upcoming changes will
use its AST level functionality.

Refs #641