Accept arrays when creating resources
All checks are now applied to every instance of a resource array.

Expiration is no longer caught early. Instead it’s processed with other
validations.

If an array of resources is uploaded, an array of resources is returned.
If a single resource is uploaded, it’s returned as it always was.
Because `multipart/form-data` doesn’t make a distinction between a
single value and an array, it’s treated as a single resource when only
one resource is present. This is consistent with the old behaviour.

Refs #651