Remove title from flow page steps if name is undefined
Right now when a flow page’s step does not have a name set, it will try to load a translation using the format of `{pages.X.steps.Y}`.
If this can’t be found, it would instead render the ID as a placeholder.

Instead of displaying that ID in the title, don’t render any title.