Fix regression in mailer
If the `smtpFrom` value is invalid, it’s parsed as `null`. This means
it’s unsafe to access the `address` field. Optional chaining is used to
fix this.

Also the logging showed the `from` function parameter, not the processed
`From` header that was used. The `From` header is now logged instead to
aid in debugging email related issues.

Some SMTP related secrets in our Kubernetes environments had a trailing
newline. `nodemailer` strips them, but that doesn’t make them valid.
They have been updated.