Take Monaco marker severity in consideration
YAML syntax errors are errors, schema violations are warnings. Custom
validation errors are now also marked as warnings for a more consistent
API.

When rendering the markers in the problems pane, the severity is also
used to render a different icon and icon color. Hint severity
diagnostics are ignored (i.e. unused anchors).