Automate login flow for CLI
Now if a user runs `appsemble login`, a browser opens on the correct
URL. If the user saves the client credentials, the CLI will receive the
client credentials from Appsemble Studio in the background.

This also fixes a security issue where Appsemble Studio could leak
access tokens if it would make requests to third parties.

Closes #582