feat: added layout option to message action
Currently every message appears at the same place in the bottom.
Depending on how an app is layed out this might not always
be very convenient.
For example, if you chose to have your navbar in the bottom
the messages block it entirely.
This commit provides this as an option for change the position
of a message on the screen.

Fixes #919