Enhance documentation
This adds `unified-consistency` as a linting check. This ensures a
consistent styling better than `remark-prettier`, because it also
validates the styling is consistent after applying other remark plugins.

Secondly this adds a table of contents to most projects. This is
verified and updated automatically using `remark-toc`.

Also strong tags were removed from headings.