Set max number of combined Playwright test failures to 6
When an issue arises with the Playwright tests in the pipeline it often takes
very long for the job finish (90 minutes). The only way to get the test artifacts
is after the job is done which makes you wait very long. This reduces it down to
around a maximum of 10 minutes.