Fix sharp import in production
Sharp is required to be imported in the root vitest.config.ts, because
sharp needs to be imported before any child processes start. However
adding sharp to the root package.json results in an import error in
the docker container. So therefore also the .eslintrc config file has
been changed to ignore any faulty import statements.