fix(server): do not decompress request action proxy responses
Axios by default decompresses all responses see
https://axios-http.com/docs/req_config

With the request action proxy endpoint it kept decompressing responses
before sending them to the client. This would result in the client
thinking the response body being the length of the compressed body which
would result in malformed responses.

This should not only fix responses that have the content-encoding header
set to gzip (and transfer-encoding not to chunked), but also improve
server performance.