Fix migration for release version 0.22.0
This migration did not work because the subscribed column already exists
on the User table (in production). Further problems would have arisen no
matter what, as the column is of type boolean and the bulk update being
applied would have failed, since it tried to update with a date value.